Split off connections further
